refactor(envex): extract type definitions into co-located types.ts files
Move inline interface/type definitions into dedicated types.ts files
within each module directory for consistent separation of concerns.
Consolidate the identical EnvScriptProps and InlineEnvScriptProps into
a single shared definition in nextjs/types.ts, re-exported under their
respective names.
