+++ packages/envex/src/nextjs/EnvScript/types.ts
@@ -1,0 +1,1 @@
+export type { EnvScriptProps } from '../types'
