+++ packages/envex/src/nextjs/InlineEnvScript/types.ts
@@ -1,0 +1,1 @@
+export type { EnvScriptProps as InlineEnvScriptProps } from '../types'
