+++ packages/envex/src/nextjs/types.ts
@@ -1,0 +1,5 @@
+import type { StandardSchemaV1 } from '@standard-schema/spec'
+
+export interface EnvScriptProps {
+  schema?: StandardSchemaV1
+}
