+++ packages/envex/src/nextjs/utils/getEnv/types.ts
@@ -1,0 +1,7 @@
+import type { StandardSchemaV1 } from '@standard-schema/spec'
+
+export interface GetEnvOptions<
+  TSchema extends StandardSchemaV1 | undefined = undefined,
+> {
+  schema?: TSchema
+}
