+++ packages/envex/src/nextjs/utils/getPublicEnv/types.ts
@@ -1,0 +1,8 @@
+import type { StandardSchemaV1 } from '@standard-schema/spec'
+
+export interface GetPublicEnvOptions<
+  TSchema extends StandardSchemaV1 | undefined = undefined,
+> {
+  schema?: TSchema
+  prefix?: string | null
+}
