+++ packages/envex/src/react/EnvexProvider/utils/fetchEnvCache/types.ts
@@ -1,0 +1,6 @@
+import type { Env } from '../../../../types'
+
+export interface FetchEnvCache {
+  promise: Promise<Env> | null
+  endpoint: string | null
+}
